refactor: move current footer to FullFooter component
